click-related cleanups
- remove some useless casts
- the type for milliseconds is now "long" to match lua integers and the field in
   struct kevent

+++ src/click_thread.h
@@ -5,5 +5,5 @@
 void click_thread_init(void);
 void click_thread_deinit(void);
 
-_Bool click_thread_submit_click(unsigned int ms, uintptr_t *id_out);
+_Bool click_thread_submit_click(long ms, uintptr_t *id_out);
 _Bool click_thread_cancel_click(uintptr_t id);
